requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
watchdog==3.0.0
GitPython==3.1.40
PyGithub==2.1.1
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
cryptography==41.0.7
python-multipart==0.0.6
